style updates, linting, moving delete to own tab
