+++ blocks/site-details/renderSiteOverview.js
@@ -4,7 +4,7 @@ import {
   safeText,
 } from '../../scripts/scripts.js';
 import renderSkeleton from '../../scripts/skeletons.js';
-import { alertDialog, confirmDialog, createDialog } from '../../scripts/dialogs.js';
+import { alertDialog, createDialog } from '../../scripts/dialogs.js';
 import renderCheckList from './renderCheckList.js';
 
 export default async function renderSiteOverview({
@@ -78,12 +78,7 @@ export default async function renderSiteOverview({
 
   <div class="checklist-container">
   </div>
-
-  <div class="danger-zone">
-    <strong>Danger zone</strong>
-    <p>Delete this project. Once you delete a project, there is no going back. Please be certain.</p>
-    <button id="delete-site-button" title="Delete your Project" class="button delete action destructive">Delete</button>
-  </div>`;
+  `;
 
   const checklistContainer = container.querySelector('.checklist-container');
   const descriptionSpan = container.querySelector('.project-description.card .project-description.span');
@@ -142,29 +137,5 @@ export default async function renderSiteOverview({
 
       dialog.setLoading(false);
     };
-  };
-
-  // MARK: delete site
-  container.querySelector('.delete').onclick = async () => {
-    window?.zaraz?.track('click site delete');
-    const block = container.closest('.site-details.block');
-
-    block.classList.add('is-deleting');
-    if (await confirmDialog('Are you sure you want to delete your site? (This can\'t be undone)')) {
-      window?.zaraz?.track('click site delete submit');
-
-      const reqDelete = await fetch(`${SCRIPT_API}/${projectDetails.darkAlleyProject ? 'da-' : ''}delete/${projectDetails.projectSlug}`, {
-        method: 'DELETE',
-        headers: { authorization: `bearer ${token}` },
-      }).catch(() => null);
-      if (reqDelete?.ok) {
-        window.location.href = '/dashboard/sites';
-      } else {
-        await alertDialog(OOPS);
-        block.classList.remove('is-deleting');
-      }
-    } else {
-      block.classList.remove('is-deleting');
-    }
   };
 }
