+++ blocks/site-details/site-details.js
@@ -8,7 +8,7 @@ import {
 import renderCampaignsAnalytics from './renderCampaignsAnalytics.js';
 import renderCampaignsAudience from './renderCampaignsAudience.js';
 import renderCampaignsOverview from './renderCampaignsOverview.js';
-import renderCheckList from './renderCheckList.js';
+import renderDangerZone from './renderDangerZone.js';
 import renderSettingsGeneral from './renderSettingsGeneral.js';
 import renderSiteAnalytics from './renderSiteAnalytics.js';
 import renderSiteOverview from './renderSiteOverview.js';
@@ -66,12 +66,6 @@ export default async function decorate(block) {
         iconSrc: '/icons/template.svg',
         renderTab: renderSiteOverview,
       },
-      // {
-      //   name: 'Checklist',
-      //   href: `${pathname}/checklist`,
-      //   iconSrc: '/icons/checklist.svg',
-      //   renderTab: renderCheckList,
-      // },
       {
         name: 'Pages',
         href: `${pathname}/pages`,
@@ -128,6 +122,12 @@ export default async function decorate(block) {
         iconSrc: '/icons/palette.svg',
         isLink: true,
         target: '_blank',
+      },
+      {
+        name: 'Delete',
+        href: `${pathname}/danger-zone`,
+        iconSrc: '/icons/delete.svg',
+        renderTab: renderDangerZone,
       },
     ],
   });
